+++ tencentcloud/iotexplorer/v20190423/errors.go
@@ -62,6 +62,9 @@ const (
 	// DB操作错误。
 	INTERNALERROR_DBOPERTAIONERROR = "InternalError.DBOpertaionError"
 
+	// 内部LoRaServer错误。
+	INTERNALERROR_INTERNALLORASERVERERROR = "InternalError.InternalLoRaServerError"
+
 	// 内部RPC错误。
 	INTERNALERROR_INTERNALRPCERROR = "InternalError.InternalRPCError"