+++ tencentcloud/lighthouse/v20200324/models.go
@@ -1671,7 +1671,18 @@ func (r *DescribeFirewallRulesTemplateResponse) FromJsonString(s string) error {
 type DescribeGeneralResourceQuotasRequest struct {
 	*tchttp.BaseRequest
 
-	// 资源名列表，取值为：USER_KEY_PAIR、GENERAL_BUNDLE_INSTANCE、STORAGE_BUNDLE_INSTANCE、ENTERPRISE_BUNDLE_INSTANCE、EXCLUSIVE_BUNDLE_INSTANCE、BEFAST_BUNDLE_INSTANCE、SNAPSHOT、BLUEPRINT、FREE_BLUEPRINT、DATA_DISK、ATTACHED_DATA_DISK、FIREWALL_RULE。
+	// 资源名列表，可取值:
+	// - GENERAL_BUNDLE_INSTANCE 通用型套餐实例
+	// - STORAGE_BUNDLE_INSTANCE 存储型套餐实例 
+	// - ENTERPRISE_BUNDLE_INSTANCE 企业型套餐实例 
+	// - EXCLUSIVE_BUNDLE_INSTANCE 专属型套餐实例
+	// - BEFAST_BUNDLE_INSTANCE 蜂驰型套餐实例
+	// - USER_KEY_PAIR 密钥对
+	// - SNAPSHOT 快照
+	// - BLUEPRINT 自定义镜像
+	// - FREE_BLUEPRINT 免费自定义镜像
+	// - DATA_DISK 数据盘
+	// - FIREWALL_RULE 防火墙规则
 	ResourceNames []*string `json:"ResourceNames,omitempty" name:"ResourceNames"`
 }