+++ tencentcloud/tsf/v20180326/errors.go
@@ -626,6 +626,9 @@ const (
 	// 重装系统，无效的镜像id。
 	INVALIDPARAMETERVALUE_INSTANCEINVALIDIMAGE = "InvalidParameterValue.InstanceInvalidImage"
 
+	// 参数 %s 取值错误。
+	INVALIDPARAMETERVALUE_INVALIDPARAMETER = "InvalidParameterValue.InvalidParameter"
+
 	// 参数格式异常。
 	INVALIDPARAMETERVALUE_INVALIDPARAMETERFORMAT = "InvalidParameterValue.InvalidParameterFormat"